feat(Client) adding a search function

+++ src/main/java/dev/felipe/clientmanagement/dto/client/ClientResponseDTO.java
@@ -4,6 +4,6 @@ import java.util.List;
 
 public record ClientResponseDTO(
         List<ClientResponseItemsDTO> clients,
-        int total
+        long total
 ) {
 }
